machines: Add a new callMachines function
This is particularly needed for the Labtops and it allows to call a full
attribute set of machines to be incorporated into a single file as one
big attribute set.

Its functionality is kind of similar to the NixOps network expressions
by providing default hostnames (in our case with a priority of 900 to
still make it overridable without pain).

Signed-off-by: aszlig <aszlig@redmoonstudios.org>