machines/labtops: Switch to use callMachines
Should make the configurations much more easy to read.

I did a small comparison between the machine derivations of the old
Labtop and the Labtop with the new callMachines implementation using:

diff -U 0 =(nix-store -qR old_labtop.drv | sort -t- -k 2) \
          =(nix-store -qR new_labtop.drv | sort -t- -k 2)

The following store paths were different in the output:

  /nix/store/...-etc.drv
  /nix/store/...-initrd.drv
  /nix/store/...-kernel-modules-shrunk.drv
  /nix/store/...-nixos.conf.drv
  /nix/store/...-nixos-system-labtop-16.09pre82222.fc92bbf-vuizvui.drv
  /nix/store/...-stage-1-init.sh.drv
  /nix/store/...-system-units.drv
  /nix/store/...-unit-systemd-modules-load.service.drv

This is okay and is due to the reversed module evaluation order, because
we now have the module definition enabling the Labtop profile in
extraConfig instead of in the root config.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>
Cc: @Profpatsch