lib/call-network: Pass nodes attribute to modules
On NixOps as well as for NixOS test networks, there is a nodes attribute
which allows to reference configuration definitions from other modules.

Currently this isn't used at all but might come in handy if we want to
cross-reference machine configurations (for example referencing IP
addresses in a static network configuration).

Signed-off-by: aszlig <aszlig@redmoonstudios.org>