lib/call-machine: Get rid of extraConfig attribute
Right now we're not using the system attribute at all and we can still
use nixpkgs.system to set the attribute for a particular machine.

So we now can pass configuration attributes to the second argument of
callMachine *directly* instead of using specific subattributes, which I
think feels is a more natural way so users don't need to look up that
"extraConfig" is for adding configuration values.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>