aszlig/profiles/base: Use IcedTea for FireFox
I use FireFox occasionally for whenever I'm forced to run ugly Java
plugins (in particular Hetzner's Lara machines use it).

I hope IcedTea is capable of running the Lara plugin, but even when not,
it's not a big deal-breaker as I didn't have to use Lara consoles for a
whole while now. And if it's the case, I can still look for a "fix" :-)

So another step towards being free of proprietary software.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>