tyree: Make the shell less annoying to use
Having the default locale set to de_DE it feels pretty weird getting
output that sounds awkward (well, obviously because German IS awkward)
and I often have to pause for a minute to understand what the message
should actually mean.

The main user of the system is rarely working in the shell, but even
during those occasions it's difficult for me if I get something like
"There is an error message saying <some blurb in German>".

Also, having command line completion is a good idea as well.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>