pkgs/tomahawk: Fix build by pinning to Qt 5.5
Right now we haven't packaged QtWebEngine for Qt 5.6, so we can't easily
switch towards using the qt5 attribute, which is Qt 5.6 already.

So we need to make sure that attica, extra-cmake-modules and qca are
using qt55 explicitly as well so we don't have any reference to Qt 5.6
anymore left (checked using nix-store -qR on the .drv).

Signed-off-by: aszlig <aszlig@redmoonstudios.org>