tests: Remove dummy test for i3
We already have an "i3wm" test in upstream <nixpkgs> which is much more
thorough than the unfinished test I've made here.

The intention of this test however was to specifically test the Vuizvui
service module. Nevertheless, it's still just a dummy test and the
"i3wm" test works much better, so let's remove it until we have a more
complete implementation.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>