pkgs/openlab/gitit: Build on Hydra
Instead of setting meta.platforms, we're setting just "platforms" here,
because the cabal build expression passes it through as meta.platforms.

Note that I'm currently setting the platforms attribute to just
x86_64-linux for now, because that is the architecture the workaround
machine "website-vm" so far has been built for.

Of course, if we'd not set the platforms at all, it would include all
supported architectures by default.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>
Cc: @Profpatsch