machines: Remove machine "website-vm"
According to @Profpatsch the whole point of this dummy machine was that
the patched gitit version should be built on Hydra.

We don't need to have such workarounds, because we're already recursing
through all packages in the Vuizvui namespace whether meta.platforms
includes a system that we support on our Hydra.

This has been done with a4d6395 so "website-vm" is obsolete now.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>