release.nix: Add vuizvui pkgs to generic channel
This adds all the packages that are marked to be built on Hydra to the
constituents of the generic channel so that we can ensure that the
channel always stays with succeeding builds.

It's especially useful for the patched gitlab that is used for the
OpenLab website VM, because it is not a NixOS system. We can simply use
the generic channel over there and stay up-to-date with it without
getting broken builds.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>
Cc: @Profpatsch