pkgs/games/humblebundle: Add Brigador
Just tested it a bit but haven't really played it (and probably won't
have time to during the next weeks).

This is the version that got released on 2016-06-02 and it needs a small
preload wrapper because it tries to load assets and other stuff relative
to the current directory. The preloader makes sure that the correct
store path is read instead.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>