taalo-build: Fix picking up derivation paths
Regression introduced by fba5c2469bd5e95857c4ffd9db3dd0529f2485f3.

The derivation paths returned by nix-instantiate are usually valid store
paths, except when used with --add-root. In case of the latter, we get
the root symlink back, so we need to do an additional readlink on it.

We also now pass -t to mktemp so that $TMPDIR or /tmp is used instead of
the current working directory.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>