pkgs/gitit: Make sure hydraPlatforms are set
Within mapTestOn, meta.hydraPlatforms attribute has precedence over
meta.platforms. The original gitit derivation has hydraPlatforms set to
an empty list and it's aliased to meta.platforms.

Unfortunately that alias doesn't work with overrideCabal, because we'd
need to override the attributes passed to the callPackage'd function
(the mentioned alias is done during function attrset unpacking).

This fixes building gitit within the generic channel as well as
including it in the pkgs.* Hydra jobs.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>
Cc: @Profpatsch