pkgs/axbo: Update to latest development branch
This one explicitly includes a MIT license, so let's actually prefer
this branch instead of the master brach.

I've also included commons and streamflyer to the list of bundled JARs
that are needed from the upstream repository. Of course, in the long
term it would be a good idea to actually package those in <nixpkgs> so
it's just temporary.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>