gajim: Add python-axolotl as a runtime dependency
This is needed to run the OMEMO plugin:

https://github.com/omemo/gajim-omemo

I'm using drvAttrs directly here, so that we can pass
propagatedBuildInputs properly to the wrapper (without duplicating the
dependencies all over the place).

Ideally, this will be fixed in <nixpkgs> but in a much more fine-grained
way in that we are going to have Nix expressions for every single plugin
and its dependencies.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>