machines/tyree: Avoid renaming network interfaces
This seems to be related to this bug:

https://bugs.launchpad.net/ubuntu/+source/network-manager/+bug/1434986

The problem here is a race condition, because NetworkManager is trying
to pick up the device _before_ the renaming and then fails to correctly
identify it.

Another solution would be to reorder the service unit after the networkd
service, but that would only work well if the interface is plugged in at
boot time and if it's pluggen in/out afterwards we likely have the same
problem again.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>