i3/conky: Suppress output to stderr
i3bar seems to close the FD for stderr, so conky will get a broken pipe
error when writing error messages to stderr.

Of course in the long run it's better to log these errors somewhere
where they can be inspected later.

But for now, it's better than terminating i3bar altogether, because
that's really annoying ;-)

Signed-off-by: aszlig <aszlig@redmoonstudios.org>