hardware/t100ha: Fix vblank issue early at initrd
Using xset for DPMS control only works as soon as we have an X session,
but we want to work around the issue as early as possible so let's set
it in preDeviceCommands.

The program here directly uses the ioctls to blank/unblank the current
console without enums/constants (14 is blank and 4 is unblank) because
it's just a workaround and supposed to go away in the near future.

If we'd wanted to use the constants from tiocl.h, we'd have to add
dependencies on Linux headers, but I'm lazy right now ;-)

Signed-off-by: aszlig <aszlig@redmoonstudios.org>