tests: Fix and improve Starbound test
Since version 1.0, the coordinates for the menu labels no longer apply
and need to be fixed. Also we no longer land on our ship but in the
protectorate building, so there won't be a quest dialog to close.

This also simplifies the test because we can now detect whether we're
in-game using OCR matching parts of the quest marker for "Attend your
graduation ceremony".

I've also increased the available memory for the server, because it
seems that for this simple test the base memory required for running a
Starbound server seems to have increased.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>