modules/starbound: Remove clear* options
These options are only a one-off setting that's causing the player and
universe files to be cleared on server startup. The service populates
all options declaratively while this option is something that IMHO
doesn't fall into this category, not to mention I fail to see why this
would have an advantage over deleting those files using the command
line.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>