pkgs/santander: Suppress substitution failure
Since NixOS/nixpkgs@b6c204f substituting a file which doesn't exist will
fail and the preFixup phase tries to substitute something in the
"install-dependency" script, which doesn't exist here in vuizvui (we're
using a later version directly from upstream Git).

I'm suppressing this simply by null'ing out the preFixup phase instead
of fixing this properly because I'm not going to stay at this bank for a
very long time.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>