machines/tishtushi: Add BFQ patch replacing CFQ
So far until version 4.7.0 I've used the BFQ patch that adds BFQ as
another I/O scheduler to the kernel. I'm using BFQ as the default
scheduler on this machine, so it makes sense to completely replace the
CFQ scheduler with the BFQ implementation.

Paolo Valente has already ported this to 4.8 according to this post:

https://groups.google.com/d/msg/bfq-iosched/RQNkbDf9rAg/DqgVDGs7AAAJ

So let's just switch to the implementation that might get mainlined
someday.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>