hardware/t100ha: Fix support for SDIO WiFi card
While I had terse Internet connectivity these days, I haven't checked
T100HA hardware related news/patchs very often.

Meanwhile a small patch by @tagorereddy popped up on hadess/rtl8723bs#80
which works around the SDIO detection.

On some devices the BIOS has an option to map the SDIO controller to PCI
instead of ACPI, but the T100HA doesn't have this option.

IIUC the issue here is that the controller is already being attached via
PCI device ID 80862280, which prevented the platform device from being
attached via ACPI.

The patch circumvents that by removing PCI device ID 80862280 while
probing for ACPI devices in SDHCI (yes, very hacky but works) and forces
the ACPI platform device to be initialized if the HID is 80860F14:02.

I've rebased and tidied up the original patch a bit to prevent warnings
during build.

Other than this patch, we now need to add the right firmware and the
corresponding parameters to hardware.firmware, so that the actual driver
for the WiFi card can be loaded.

The firmware parameters are stored in EFI on the host itself:

/sys/firmware/efi/efivars/nvram-74b00bd9-805a-4d61-b51f-43268123d113

So I added a copy of it to prevent impurities.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>