pkgs/santander: Trim down closure size
Since the version bump of wineStaging, stuff like mono and gecko are
included when creating a new wineprefix. We can avoid this by overriding
DLLs using "mscoree,mshtml=".

In addition to that, we now use wineMinimal instead of the wineStaging
attribute from nixpkgs so that we really have as few options compiled in
as possible, especially because we really don't need them for the
plugin.

This trims down the closure size by a few hundred megabytes.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>