pkgs/grandpa: Remove dependency on bsddb
It seems (haven't really tested) that since
NixOS/nixpkgs@bee439207bd729eee0bf5dcae5be44320cee45cd using bsddb is no
longer necessary and built in into Python.

I haven't tested whether this is true or not, but the attribute is gone
so let's make sure that the expression evaluates for now.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>