machines/fliewatuet: Fix evaluation error
The jackline package only existed in nixpkgs for a short period of time
until it got reverted in
NixOS/nixpkgs@7e0ca08bfa19304b446f860a8899122ef019c51a.

As soon as jackline enters nixpkgs again, this commit can either be
reverted or just left as-is, because it uses pkgs.jackline only if the
attribute exists.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>
Cc: @sternenseemann