modules/git: Fix/flesh out type for Git config
We need to properly separate attrsOf from the nested attrsOf because
otherwise the either type isn't able to distinguish between both types
during merging of the subtypes.

Now we have sub sections which is '[section "subsection"]' in Git's
config syntax and attrsOf (attrsOf ...) on our side and simple sections
like '[section]' which is just a plain attrsOf. This way we have an
additional type constraint in that we can't mix up subsections with
sections by the same name.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>