modules/gnupg: Support --supervised since 2.1.16
See the release announcement at:

https://lists.gnupg.org/pipermail/gnupg-announce/2016q4/000398.html

Unfortunately we still need the wrapper, because we need to pick up the
PID of the socket endpoint in order to gather various information we can
pass to pinentry (which then for example can recognize that the actual
SSH client is using X or is using a particular TTY).

On the upside however, this is a step into a direction I didn't imagine
to happen anytime soon, given the following statement from Werner Koch
back then:

https://lists.gnupg.org/pipermail/gnupg-devel/2014-November/029104.html

However I don't think the way I'm picking up the PID and doing
inspection of /proc/PID is going to happen in upstream GnuPG anytime
soon. But after cleaning up and doing it as a patch I might now consider
upstreaming it.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>