machines/dnyarri: Enable sane with hplip
This machine is used for collecting all that useless paperwork gathered
since a few years, so I need scanner support.

Unfortunately the scanner (it's part of a HP Officejet all-in-one
thingy) needs to have a proprietary plugin for hplip in order to work,
which is not nice and I'm not very proud about needing to do this.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>