aszlig/kernel: Update BFQ to v8r7 for 4.10-rc1
The patch is from https://github.com/linusw/linux-bfq/tree/bfq-v8.

It's a combined patch until the parent of the branch's head, because the
latest commit is a work-in-progress commit.

I have only tested evaluation and didn't test whether the patch actually
applies yet, because the machines currently using the BFQ patch are
broken because the old BFQ patch no longer applies for kernel 4.10.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>