aszlig/kernel: Fix up BFQ patch for kernel 4.10
First of all, using .patch from the GitHub compare view is not going to
work because the concatenated diffs are in reverse order for use within
"git apply".

And the second thing why it's not working is that the patch has a hunk
that changes the version in Makefile to add an extra version -rc1-bfq,
which would only apply for kernel 4.10-rc1 but not for subsequent
release canidates.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>