core/tests: Add a few missing callSubTests tests
Haha, so that's what I got from introducing callSubTests in
NixOS/nixpkgs@b3337edd0b77f04d0abe4a29642a5e4fdb4ad692.

While in general this is an improvement, it makes it harder for vuizvui
to gather tests that were added or removed.

Maybe we should write an updater for this soon.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>