pkgs/santander: Whitelist webkitgtk2 for dwb
WebKitGTK+ has a bunch of security vulnerabilities:

  * WSA-2017-0002
  * WSA-2017-0001
  * WSA-2016-0006
  * WSA-2016-0005
  * WSA-2016-0004

We're using dwb in conjunction with this insecure version only for *one*
specific application with *one* specific web site, so the attack surface
is relatively low.

The main issue with this is that the plugin required for the Santander
page is based on NPAPI and most of the browsers out there do not support
NPAPI, except a few ones:

https://en.wikipedia.org/wiki/NPAPI#Browser_support

I've tried to embed the plugin within Midori but I couldn't get it to
work so far (Midori just says "No plugin available"), so I'm sticking
with dwb for now, because the other browsers listed here either have the
same problem (uzbl) or are simply too heavyweight.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>