taalo-build: Use Nix 1.11 for the backend
Nix 1.12 already got rid of most of the Perl stuff so the Perl script
backing taalo-build doesn't work anymore because Nix/CopyClosure.pm is
no longer available.

One way to fix this properly is to part the backend to C++ and use the
new ssh-ng store backend. This would also have the advantage that we
don't need to hardcode protocol internals here anymore.

Unfortunately, Nix 1.12 isn't stable yet and we would break Nix 1.11 if
we use the new store API.

So that's why I'm pinning the backend to use Nix 1.11 instead and leave
that ugly Perl script until Nix 1.2 is stable.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>