core/common: Add Headcounter Hydra to binaryCaches
This makes sense to put into the core module here, because we already
have the public key in there as well and if we want to use the binary
cache outside of a channel it's a good idea to provide it here by
default.

Being a default of course also means that users can easily override this
using:

  nix.binaryCaches = mkForce [ "something else" ];

I've also removed the redundant setting of that option in all of the
machine configurations.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>
Cc: @sternenseemann