pkgs: Improve callPackageScope customisation
By default callPackageWith adds an overrideDerivation attribute, but
that won't work with our new package scopes, so we add an override
attribute by ourselves without the overrideDerivation attribute.

That aside, we now use functionArgs on not only the superset of packages
but also for the scope utility functions (callPackage/callPackage_i686)
that we pass down to our new package scope. If we didn't do that, the
Nix expression of the subscope would always need to define callPackage
and callPackage_i686 in their function arguments, regardless of whether
it's needed or not.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>