starbound: Update to version 1.3.0
Major upstream changes are:

 * Navigation Overhaul: The cockpit navigation interface has been
   completely overhauled. Planets, moons and ships are all visible and
   orbit each other in real time. Systems will now contain NPC stations,
   friendly and hostile NPC ships, and strange space anomalies to visit!

 * Customizable Mechs: Explore these new space locations in customizable
   spacefaring mechs! Traverse hostile space in zero gravity, fight
   powerful new space monsters, and collect unique rewards as you
   upgrade your mech to progress through more difficult hazards. Mechs
   can also be deployed to planets, to crush your enemies with
   overwhelming firepower!

 * Modular Space Stations: Make a permanent home among the stars with
   player-owned stations! Use a station transponder to place your
   station into orbit, then expand it with modular rooms to suit your
   needs.

The full changes can be found at the announcement blog post at:

http://playstarbound.com/spacefarer-update/

One of the changes not listed there is that the archive now consists of
a server_linux and client_linux directory, where the latter is
structured the same as in previous versions. However, both contain the
server binary and both of these binaries match in content.

So I'm assuming that the server_linux directory is only a trimmed-down
version in terms of assets but otherwise pretty much the same.

I've also fixed the VM test, which didn't recognize the font of
"Species" anymore, so we're now matching on "randomise".

In addition to that I've added a sleep of 30 seconds before the final
screenshot, so we get a picture of the fully rendered intro scenario.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>