pkgs: Remove Tomahawk from the source tree
Tomahawk is no longer actively developed and the current state within
vuizvui is also broken. So after asking the users of brawndo and tyree
whether they still use it they answered with a "no", so it doesn't make
sense to fix up that package if noone is using it anyway.

This has also been announced via the README in the official GitHub
repository, where the change + comments can be seen here:

https://github.com/tomahawk-player/tomahawk/commit/c8389592488c07079

Signed-off-by: aszlig <aszlig@redmoonstudios.org>