modules/aszlig: Cope with new xrandrHeads options
The way xrandrHeads are addressed has changed with the following
upstream merge commit:

  NixOS/nixpkgs@9dca737d6269759745c9e68ea462446cde4d9be9

Now xrandrHeads are no longer a list of strings but a list of attribute
sets, but we new do have even more information about a particular head.

So apart from fixing this, we new also set --primary whenever applicable
in "xreset".

Signed-off-by: aszlig <aszlig@redmoonstudios.org>