devhell: Mergo overrides/pkgs into vuizvui
This introduces another profile module called "packages", which contains
all the package configuration (including overrides) of all the machines
in the devhell namespace.

The machine-specific configuration is now merged into the machine
configurations the same way as we've done previously with the services.

One major difference here is that the haskellPackages workaround is no
longer needed in the package configuration, as it is handled by vuizvui.

Tested this by evaluating all machines and all evaluations succeeded.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>
Cc: @devhell