modules/kernel: Really enable BFQ by default
I previously wrote that patch in a hurry, so I thought it would be
enough to set CONFIG_DEFAULT_IOSCHED to "bfq". But in block/elevator.c
the actual default for blk-mq is a constant and can't be configured via
CONFIG_DEFAULT_IOSCHED.

So we're now patching just that constant and nothing more.

Also, I've enabled CONFIG_DM_MQ_DEFAULT, because the DM devices need to
be switched to blk-mq as well and for example on dnyarri I'm actually
using the device mapper for LUKS.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>