Remove service and tests for multipath-vpn
The service and test has been broken for a long time now and nobody
really has any interest in using it or even fixing it, so I'm removing
it to decrease the amount of crap we have in there.

If somebody still wants to use this someday we can still bring it back.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>