Convert packageOverrides to overlays
There are two other occasions where packageOverrides are still in use,
one of them is @sternenseemann's fliewatuet machine and another one is
@devhell's package profile.

I've replaced every other occurence of packageOverrides and replaced it
with overlays and checked the store path hash after evaluation.

The reason why I left @sternenseemann's fliewatuet alone was that for
I wasn't quite sure whether "bluez = pkgs.bluez5" is still needed or
intentional to pin it to version 5. Because if it's not the case the
packageOverrides can just be dropped because bluez is already bluez5 in
upstream <nixpkgs>.

For @devhell's package profile, I did the conversion, but the store path
turned out to be a different one than what it was prior to the change.

I did take a quick look at the requisites of the drv and found that sox
was one of the different paths that led to the change in the final store
path.

This obviously needs to be tested and/or investigated first.

Other than that, the main reason why I'm moving everything to overlays
is that it's the replacement for packageOverrides and also has a better
way to pass through chains of overrides than packageOverrides had.

My guess is that after NixOS 17.09, the old packageOverrides function
will be removed, so let's make sure we're ready for that.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>
Cc: @devhell, @sternenseemann